% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zip_data.R
\name{add_zip_data}
\alias{add_zip_data}
\title{Add ZIP Code related Info to data frame}
\usage{
add_zip_data(df)
}
\arguments{
\item{df}{\code{tibble} returned from \code{download_dataset()}}
}
\value{
A \code{tibble}
}
\description{
Add ZIP Code related Info to data frame
}
\examples{
search_datasets(specialty = "vascular surgery") |>
download_dataset() |>
dplyr::slice_sample(n = 5) |>
add_zip_data()
}
