% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costoffice-package.R
\docType{package}
\name{costoffice-package}
\alias{costoffice}
\alias{costoffice-package}
\title{costoffice: Download Physician Office Visit Cost Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides various functions to enable downloading CMS physician office visit cost data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andrewallenbruce/costoffice}
  \item \url{https://andrewallenbruce.github.io/costoffice/}
  \item Report bugs at \url{https://github.com/andrewallenbruce/costoffice/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrewbruce.himni@gmail.com} [copyright holder]

}
\keyword{internal}
