% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_types}
\alias{df_types}
\title{Returns a tibble describing the structure of a data frame}
\source{
https://r4ds.hadley.nz/iteration.html#reading-multiple-files
}
\usage{
df_types(df)
}
\arguments{
\item{df}{data frame}
}
\description{
Returns a tibble describing the structure of a data frame
}
\examples{
search_datasets_arrow() |>
df_types()
}
