% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset}
\alias{download_dataset}
\title{Download A Physician Office Visit Costs Data set}
\usage{
download_dataset(df)
}
\arguments{
\item{df}{tidytable from \code{search_datasets()}}
}
\value{
A \code{tidytable}
}
\description{
A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
search_datasets(specialty = "vascular surgery") |>
download_dataset() |>
tidytable::slice_head(n = 10)
}
}
