% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_dataset_arrow}
\alias{download_dataset_arrow}
\title{Download A Physician Office Visit Costs Data set}
\usage{
download_dataset_arrow(specialty)
}
\arguments{
\item{specialty}{physician specialty}
}
\value{
An \code{arrow} Table object
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
search_datasets(specialty = "addiction medicine") |>
download_dataset_arrow()
}
}
