% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_datasets}
\alias{download_datasets}
\title{Download A Physician Office Visit Costs Data set}
\usage{
download_datasets(n)
}
\arguments{
\item{n}{number of datasets to download, number between 1-83}
}
\value{
An \code{arrow} Table object
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
download_datasets(n = 2)
}
}
