% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search Physician Office Visit Costs Data sets}
\source{
Centers for Medicare & Medicaid Services
}
\usage{
search_datasets(specialty = NULL)
}
\arguments{
\item{specialty}{Provider's medical specialty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
}
\details{
The Pending Initial Logging and Tracking (L & T) data set provides
a list of pending applications that have not been processed by CMS
contractors.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\note{
Update Frequency: \strong{Weekly}
}
\examples{
#search_datasets()
}
