% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zipcoder.R
\name{use_zipcoder}
\alias{use_zipcoder}
\title{Add Geocoding Info to dataset via zipcodeR database}
\usage{
use_zipcoder(df)
}
\arguments{
\item{df}{tidytable from \code{download_dataset()}}
}
\value{
A \code{tidytable}
}
\description{
A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
search_datasets(specialty = "vascular surgery") |>
download_dataset() |>
tidytable::slice_tail(n = 5) |>
use_zipcoder()
}
