% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zipcoder.R
\name{use_zipcoder_arrow}
\alias{use_zipcoder_arrow}
\title{Add Geocoding Info to dataset via zipcodeR database}
\usage{
use_zipcoder_arrow(df)
}
\arguments{
\item{df}{An \code{arrow} Table object from \code{download_dataset_arrow()}}
}
\value{
An \code{arrow} Table object
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
search_datasets_arrow(specialty = "vascular surgery") |>
download_dataset_arrow() |>
use_zipcoder_arrow()
}
}
