% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_dataset.R
\name{write_datasets}
\alias{write_datasets}
\title{Download A Physician Office Visit Costs Data set}
\usage{
write_datasets(df, dir = "D:/cost_office_csvs/", ext = ".csv")
}
\arguments{
\item{df}{A returned \code{tidytable} from \code{search_datasets()}}

\item{dir}{Directory to save to}

\item{ext}{File extension, default is ".csv"}
}
\value{
A \code{tidytable} data.table
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
write_datasets(specialty = "vascular surgery",
               dir = "D:/cost_office_csvs/",
               ext = ".csv")
}
}
\seealso{
Other download: 
\code{\link{download_datasets}()}
}
\concept{download}
