% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db_funcs.R
\name{write_duckdb}
\alias{write_duckdb}
\title{Write Physician Office Visit Costs Data sets to a DuckDB database}
\usage{
write_duckdb(
  dir = "D:/cost_office_csvs/",
  db_dir = "D:/cost_office_duckdb/",
  db_name = "cost_office.duckdb"
)
}
\arguments{
\item{dir}{directory containing csvs to load into DuckDB}

\item{db_dir}{directory where database will be created}

\item{db_name}{name of database ending with .duckdb extension}
}
\value{
invisible; directory & database created
}
\description{
Write Physician Office Visit Costs Data sets to a DuckDB database
}
\examples{
\dontrun{
write_duckdb(dir = "D:/cost_office_csvs/",
         db_dir = "D:/cost_office_duckdb/",
         db_name = "cost_office.duckdb")
}
}
\concept{databases}
