% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zipcode_db.R
\docType{data}
\name{zipcode_db}
\alias{zipcode_db}
\title{Zip code database}
\format{
\subsection{\code{zipcode_db}}{

A tidytable with 41,877 rows and 10 columns:
\describe{
\item{city}{City name}
\item{county}{County name}
\item{state}{State abbreviation}
\item{zip_code}{5 character zip}
\item{lat}{latitude}
\item{lng}{longitude}
\item{bounds_west}{boundary west}
\item{bounds_east}{boundary east}
\item{bounds_north}{boundary north}
\item{bounds_south}{boundary south}
}
}
}
\usage{
zipcode_db
}
\description{
A subset of the zip code database from the \code{zipcodeR} package
}
\keyword{datasets}
