% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_dataset.R
\name{download_datasets}
\alias{download_datasets}
\title{Download A Physician Office Visit Costs Data set}
\usage{
download_datasets(
  specialty = NULL,
  keyword = NULL,
  zipcoder = TRUE,
  drop_na = TRUE
)
}
\arguments{
\item{specialty}{Search for exact medical specialty, e.g. 'cardiology'}

\item{keyword}{Search for partial matches, e.g. 'medicine'}

\item{zipcoder}{Adds information from \code{use_zipcoder()}; default is TRUE}

\item{drop_na}{Drop rows containing missing values; default is TRUE}
}
\value{
A \code{tidytable} data.table
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
download_datasets(keyword = "medicine")
}
}
\seealso{
Other download: 
\code{\link{download_dataset_purrr}()},
\code{\link{download_dataset}()}
}
\concept{download}
