% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_datasets.R
\name{search_datasets}
\alias{search_datasets}
\title{Search Physician Office Visit Costs Data sets by Specialty}
\usage{
search_datasets(specialty = NULL, keyword = NULL)
}
\arguments{
\item{specialty}{Provider's medical specialty}
}
\value{
A \code{tidytable}
}
\description{
A list of enrollment applications pending CMS contractor
review for physicians & non-physicians.
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
search_datasets()
}
