% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_zipcoder.R
\name{use_zipcoder}
\alias{use_zipcoder}
\title{Add Geocoding Info to dataset via zipcodeR database}
\usage{
use_zipcoder(df, drop_na = TRUE, full = FALSE)
}
\arguments{
\item{df}{\code{tidytable} from \code{download_dataset()}}

\item{drop_na}{Drop rows containing missing values; default is TRUE}

\item{full}{Add full zip code information (demographics, geocodes, zcta crosswalk); default is FALSE}
}
\value{
A \code{tidytable}
}
\description{
\subsection{Links}{
\itemize{
\item \href{https://data.cms.gov/provider-data/search?page-size=50&theme=Physician\%20office\%20visit\%20costs}{Physician Office Visit Costs (Data.CMS.gov)}
}
}
}
\examples{
\dontrun{
search_datasets(specialty = "vascular surgery") |>
download_dataset() |>
tidytable::slice_sample(n = 5) |>
use_zipcoder()
}
}
