% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zip_db.R
\docType{data}
\name{zip_db}
\alias{zip_db}
\title{Zip code database}
\format{
\subsection{\code{zip_db}}{

A tibble with 44,950 rows and 6 columns:
\describe{
\item{zip_code}{5-digit U.S. ZIP code}
\item{city}{Major city serving the ZIP code}
\item{county}{Name of county containing the ZIP code}
\item{state}{Two-digit state code for ZIP code location}
\item{region}{Region state belongs to for ZIP code location}
\item{division}{Total population of county related to ZIP code}
}
}
}
\usage{
zip_db
}
\description{
A subset of data from the \code{zipcodeR} & \code{usa} packages
}
\keyword{datasets}
